requirements.txt
aiofiles
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
googleapis-common-protos==1.70.0
grpcio==1.73.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
llvmlite==0.44.0
lz4==4.4.4
MarkupSafe==3.0.2
more-itertools==10.7.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.4.4
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
openai==1.86.0
openai-whisper @ git+https://github.com/openai/whisper.git@dd985ac4b90cafeef8712f2998d62c59c3e62d22
packaging==24.2
passlib[bcrypt]==1.7.4
pinecone==7.0.2
pinecone-plugin-assistant==1.7.0
pinecone-plugin-interface==0.0.7
psycopg2-binary
protobuf==5.29.5
propcache==0.3.2
protoc-gen-openapiv2==0.0.1
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.3.0
python-multipart==0.0.20
regex==2024.11.6
requests==2.32.4
setuptools==80.9.0
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
sqlalchemy==2.0.30
starlette==0.46.2
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
yarl==1.20.1


